requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-telegram-bot==20.7
requests>=2.31.0
apscheduler>=3.10.0
